refactor: modal toggle 제어위치 변경
