+++ client/src/components/practice/short/InfoBar/index.tsx
@@ -1,35 +1,20 @@
 import { Box, Flex, Image, Text } from '@chakra-ui/react';
-import { useState } from 'react';
 
-import Confirm from '@/components/common/Confirm';
 import IconButton from '@/components/common/IconButton';
-import {
-  useContextShortTyping,
-  useContextShortTypingHandler,
-  useContextTypingResultModal,
-} from '@/components/practice/short/_hook/contextShortTyping';
+import { useContextShortTyping } from '@/components/practice/short/_hook/contextShortTyping';
 import GrassEllipse from '@/components/practice/short/InfoBar/GrassEllipse';
 import InfoBarItem from '@/components/practice/short/InfoBar/InfoBarItem';
 import ModeList from '@/components/practice/short/InfoBar/ModeList';
 import { Exit } from '@/icons/Exit';
 import { getSecondToMMSSFormat } from '@/utils/time';
 
+interface InfoBarProps {
+  onExit: () => void;
+}
+
 // TODO : 짧은글, 긴글에서 공통적으로 사용됨
-export default function InfoBar() {
+export default function InfoBar({ onExit }: InfoBarProps) {
   const { time, typingAccuracy, typingCount, typingWpm } = useContextShortTyping();
-  const { timePlay, timePause } = useContextShortTypingHandler();
-  const { handleResultModalOpen } = useContextTypingResultModal();
-  const [isExitModalOpen, setIsExitModalOpen] = useState(false);
-
-  const handleExitModalOpen = () => {
-    timePause();
-    setIsExitModalOpen(true);
-  };
-
-  const handleExitModalClose = () => {
-    timePlay();
-    setIsExitModalOpen(false);
-  };
 
   return (
     <>
@@ -54,7 +39,7 @@ export default function InfoBar() {
         <Box flex={1}>
           <Flex justifyContent='space-between'>
             <ModeList />
-            <IconButton icon={<Exit />} onAction={handleExitModalOpen} />
+            <IconButton icon={<Exit />} onAction={onExit} />
           </Flex>
 
           <Flex border='1px solid rgb(0, 0, 0)' borderRadius={10} h={'56px'}>
@@ -74,15 +59,6 @@ export default function InfoBar() {
           </Flex>
         </Box>
       </Flex>
-      {/* 나가기 모달에서, 계속하기를 누르면 다시 경과시간 카운트를 증가? */}
-      <Confirm
-        header={'정말로 그만 두시겠어요?'}
-        isOpen={isExitModalOpen}
-        onClose={handleExitModalClose}
-        onAction={handleResultModalOpen}
-        actionLabel='그만하기'
-        closeLabel='계속하기'
-      />
     </>
   );
 }
